requirements.txt
matplotlib==3.7.0
mmaction2.egg==info
mmcv==1.7.0
numpy==1.25.1
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
scikit_learn==1.2.2
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.64.1
